backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0

# Google Speech-to-Text (Primary)
google-cloud-speech==2.21.0

# Local Whisper dependencies (Fallback)
openai-whisper==20231117
torch>=1.10.0
torchaudio>=0.10.0